refactor: Added useGetProgramBySlug custom hook, reused it in DetailedProgramPageContent
